Honor router root if configured
The login and logout routes does not currently take into account the configuration of an alternate `root` on the router other than the default which is undefined. Say for example that the application runs with the root `/myapp/`. Then the added routes should be e.g. `/signin-oidc` and not `/myapp/signin-oidc`.